Fix xgboost base_score for logistic loss
The default global base_score is 0.5. For classifier with logistic
loss, the base_score should be transformed into margin (which is 0.0).

Due to a mistake in the change of function signature (float vs double),
one of the function overloading falled back to default impelmentation
which does no-ops on transform probability to margin, thus causing
the base_score to be off for classification problems.

Change-Id: If7c9fea702551e0fa67243dca51928aab5addc4e
(cherry picked from commit 581afec85e524f124dfdf2e8ec420e01113bb4d2)